% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cbc_survey.R
\name{create_cbc_survey}
\alias{create_cbc_survey}
\title{Creating a Choice Based Conjoint Survey}
\usage{
create_cbc_survey(level_vec, var_names)
}
\arguments{
\item{level_vec}{A vector of variable levels}

\item{var_names}{A vector of variable names}
}
\value{
sorted_survey A sorted CBC survey
}
\description{
Takes in a vector of levels and variable names, then creates a survey
}

